+++ docs/conf.py
@@ -18,7 +18,9 @@
 # absolute, like shown here.
 #
 import os
+
 import sys
+
 sys.path.insert(0, os.path.abspath('..'))
 
 # -- General configuration ---------------------------------------------