+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/datasources/subscriptions/SubscriptionsDbConnector.groovy
@@ -1,14 +1,14 @@
 package life.qbic.portal.sampletracking.datasources.subscriptions
 
 import groovy.util.logging.Log4j2
-
+import life.qbic.business.DataSourceException
+import life.qbic.business.project.load.SubscribedProjectsDataSource
 import life.qbic.business.project.subscribe.Subscriber
 import life.qbic.business.project.subscribe.SubscriptionDataSource
-
-import life.qbic.business.DataSourceException
 import life.qbic.portal.sampletracking.datasources.database.ConnectionProvider
 
 import java.sql.Connection
+import java.sql.PreparedStatement
 import java.sql.ResultSet
 import java.sql.Statement
 
@@ -20,7 +20,7 @@ import java.sql.Statement
  * @since 1.0.0
  */
 @Log4j2
-class SubscriptionsDbConnector implements SubscriptionDataSource {
+class SubscriptionsDbConnector implements SubscriptionDataSource, SubscribedProjectsDataSource {
   
   private final ConnectionProvider connectionProvider
   
@@ -135,7 +135,7 @@ class SubscriptionsDbConnector implements SubscriptionDataSource {
           }
     }
     
-    private void removeSubscription(Connection connection, int subscriberId, String projectCode) {
+    private static void removeSubscription(Connection connection, int subscriberId, String projectCode) {
           // we do not need to check if the subscription exists, here, as removing a non-existent row does not lead to errors
           String query = "DELETE FROM subscription WHERE project_code = ? AND subscriber_id = ?"
 
@@ -170,14 +170,32 @@ class SubscriptionsDbConnector implements SubscriptionDataSource {
         String query = "SELECT id FROM subscription WHERE project_code = ? AND subscriber_id = ? "
         Connection connection = connectionProvider.connect()
         boolean isAlreadySubscribed = false
-            connection.withCloseable {
-                def statement = connection.prepareStatement(query)
-                statement.setString(1, projectCode)
-                statement.setInt(2, subscriberId)
-                ResultSet resultSet = statement.executeQuery()
-                if(resultSet.next()){
-                    isAlreadySubscribed = true}
+        connection.withCloseable {
+            def statement = connection.prepareStatement(query)
+            statement.setString(1, projectCode)
+            statement.setInt(2, subscriberId)
+            ResultSet resultSet = statement.executeQuery()
+            if (resultSet.next()) {
+                isAlreadySubscribed = true
             }
+        }
         return isAlreadySubscribed
+    }
+
+    @Override
+    List<String> findSubscribedProjectCodesFor(Subscriber subscriber) {
+        List<String> subscribedProjects = []
+        int subscriberId = fetchExistingSubscriberId(subscriber)
+        String query = "SELECT project_code FROM subscription WHERE subscriber_id = ?"
+        Connection connection = connectionProvider.connect()
+        connection.withCloseable {
+            PreparedStatement statement = connection.prepareStatement(query)
+            statement.setInt(1, subscriberId)
+            ResultSet resultSet = statement.executeQuery()
+            while(resultSet.next()) {
+                subscribedProjects <<  resultSet.getString("project_code")
+            }
+        }
+        return subscribedProjects
     }
 }